pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "aind-data-mcp"
description = "Generated from aind-library-template"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Allen Institute for Neural Dynamics"}
]
classifiers = [
    "Programming Language :: Python :: 3"
]
readme = "README.md"
dynamic = ["version"]

dependencies = [
    "aind-data-access-api",
    "zombie-squirrel>=0.15.2",
    "fastmcp",
    "hdmf-zarr>=0.11.1",
    "langchain-core>=0.3.60",
    "pydantic>=2.11.4",
    "suffix-trees>=0.3.0",
]

[project.scripts]
aind-data-mcp = "aind_data_mcp.data_access_server:main"

[project.optional-dependencies]
benchmark = [
    "strands-agents",         # agent framework (Bedrock + MCP)
    "boto3",                  # Bedrock Converse API for the judge
]

[dependency-groups]
dev = [
    'black',
    'coverage',
    'flake8',
    'interrogate',
    'isort',
    'Sphinx',
    'furo'
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "aind_data_mcp.__version__"}

[tool.black]
line-length = 79
target_version = ['py311']
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
    | .conda
  )/
  | .gitignore
)
'''

[tool.coverage.run]
omit = ["*__init__*"]
source = ["aind_data_mcp", "tests"]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
    "^from .* import .*",
    "^import .*",
    "pragma: no cover"
]
fail_under = 100

[tool.isort]
line_length = 79
profile = "black"
skip = [".conda"]

[tool.interrogate]
exclude = ["setup.py", "docs", "build", ".conda"]
fail-under = 100
